checks for report file existance, missing requires

+++ lib/rspec/rotten/example_store.rb
@@ -1,15 +1,19 @@
 require 'json'
 require 'time'
 require 'rspec/core/formatters/console_codes'
+require 'rspec/rotten/configuration'
 
 module Rspec
   module Rotten
     class ExampleStore
+      # 1 year duration when AS is not available
+      ROTTEN_DURATION = Time.parse("2016-01-01") - Time.parse("2015-01-01")
+
       attr_accessor :file_path, :records
 
       def initialize()
         @file_path = Rspec::Rotten::Configuration.results_file
-        @records = JSON.parse(read_example_data)
+        @records = File.exists?(@file_path) ? JSON.parse(read_example_data) : []
       end
 
       def find(example)
@@ -29,7 +33,8 @@ module Rspec
 
       # use config for date
       def rotten
-        @rotten ||= @records.select {|x| x['date'] < Time.now() - Rspec::Rotten::Configuration.time_to_rotten }
+        rotten_duration = Rspec::Rotten::Configuration.time_to_rotten || ROTTEN_TIME
+        @rotten ||= @records.select {|x| !x['date'].nil? && x['date'] < Time.now() - rotten_duration }
       end
 
       def notify_rotten
@@ -47,10 +52,12 @@ module Rspec
       private
 
       def read_example_data
-        f = File.open(@file_path, 'r')
-        data = f.read
-        f.close
-        data
+        if File.exists? @file_path
+          f = File.open(@file_path, 'r')
+          data = f.read
+          f.close
+          data
+        end
       end
     end
   end
